+++ crypto_payment_sync/models/__init__.py
@@ -9,3 +9,4 @@ from . import crypto_transaction_evm
 from . import payment_transaction
 from . import crypto_wallet_address_event
 from . import account_payment
+from . import accounting_support